requirements.txt
pandas>=2.2.0
numpy>=1.26.0
boto3>=1.28.0
psycopg2-binary>=2.9.9
apache-airflow>=2.7.0
great-expectations>=0.17.0
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.29.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0